requirements.txt
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
anthropic>=0.49.0
pyyaml>=6.0.1
jinja2>=3.1.4
